refactor(park-keeper): NavOverlay 千行檔拆分為三個盤面元件
- 純搬移拆出 CompassArcStage（弧形盤面＋幾何渲染）、CompassCapsule（降級膠囊）、NavPermissionSheet（權限/校準卡）
- NavOverlay 1060 → 469 行；JSX、樣式、動畫參數與 testid 逐字保留，行為零變更
- 測試斷言零刪除（44 檔 632 案原樣通過），比照 issue #733 MiniMap 拆檔紀律，收斂 #747 主體

測試：pnpm typecheck 通過；pnpm exec vitest run（44 files、632 passed）
